[Lens] remove unused and duplicated code
