+++ src/plugins/chart_expressions/expression_xy/public/helpers/annotations.tsx
@@ -5,45 +5,16 @@
  * in compliance with, at your election, the Elastic License 2.0 or the Server
  * Side Public License, v 1.
  */
+import React from 'react';
 import { Position } from '@elastic/charts';
+import { EuiFlexGroup, EuiIcon, EuiIconProps, EuiText } from '@elastic/eui';
+import classnames from 'classnames';
 import type { IconPosition, YAxisMode, YConfig } from '../../common/types';
+import { getBaseIconPlacement } from '../components';
 import { hasIcon } from './icon';
+import { annotationsIconSet } from './annotations_icon_set';
 
 export const LINES_MARKER_SIZE = 20;
-
-export const computeChartMargins = (
-  referenceLinePaddings: Partial<Record<Position, number>>,
-  labelVisibility: Partial<Record<'x' | 'yLeft' | 'yRight', boolean>>,
-  titleVisibility: Partial<Record<'x' | 'yLeft' | 'yRight', boolean>>,
-  axesMap: Record<'left' | 'right', unknown>,
-  isHorizontal: boolean
-) => {
-  const result: Partial<Record<Position, number>> = {};
-  if (!labelVisibility?.x && !titleVisibility?.x && referenceLinePaddings.bottom) {
-    const placement = isHorizontal ? mapVerticalToHorizontalPlacement('bottom') : 'bottom';
-    result[placement] = referenceLinePaddings.bottom;
-  }
-  if (
-    referenceLinePaddings.left &&
-    (isHorizontal || (!labelVisibility?.yLeft && !titleVisibility?.yLeft))
-  ) {
-    const placement = isHorizontal ? mapVerticalToHorizontalPlacement('left') : 'left';
-    result[placement] = referenceLinePaddings.left;
-  }
-  if (
-    referenceLinePaddings.right &&
-    (isHorizontal || !axesMap.right || (!labelVisibility?.yRight && !titleVisibility?.yRight))
-  ) {
-    const placement = isHorizontal ? mapVerticalToHorizontalPlacement('right') : 'right';
-    result[placement] = referenceLinePaddings.right;
-  }
-  // there's no top axis, so just check if a margin has been computed
-  if (referenceLinePaddings.top) {
-    const placement = isHorizontal ? mapVerticalToHorizontalPlacement('top') : 'top';
-    result[placement] = referenceLinePaddings.top;
-  }
-  return result;
-};
 
 // Note: it does not take into consideration whether the reference line is in view or not
 
@@ -93,36 +64,124 @@ export function mapVerticalToHorizontalPlacement(placement: Position) {
   }
 }
 
-// if there's just one axis, put it on the other one
-// otherwise use the same axis
-// this function assume the chart is vertical
-export function getBaseIconPlacement(
-  iconPosition: IconPosition | undefined,
-  axesMap?: Record<string, unknown>,
-  axisMode?: YAxisMode
-) {
-  if (iconPosition === 'auto') {
-    if (axisMode === 'bottom') {
-      return Position.Top;
-    }
-    if (axesMap) {
-      if (axisMode === 'left') {
-        return axesMap.right ? Position.Left : Position.Right;
-      }
-      return axesMap.left ? Position.Right : Position.Left;
-    }
+export function MarkerBody({
+  label,
+  isHorizontal,
+}: {
+  label: string | undefined;
+  isHorizontal: boolean;
+}) {
+  if (!label) {
+    return null;
   }
-
-  if (iconPosition === 'left') {
-    return Position.Left;
+  if (isHorizontal) {
+    return (
+      <div className="eui-textTruncate" style={{ maxWidth: LINES_MARKER_SIZE * 3 }}>
+        {label}
+      </div>
+    );
   }
-  if (iconPosition === 'right') {
-    return Position.Right;
-  }
-  if (iconPosition === 'below') {
-    return Position.Bottom;
-  }
-  return Position.Top;
+  return (
+    <div
+      className="xyDecorationRotatedWrapper"
+      style={{
+        width: LINES_MARKER_SIZE,
+      }}
+    >
+      <div
+        className="eui-textTruncate xyDecorationRotatedWrapper__label"
+        style={{
+          maxWidth: LINES_MARKER_SIZE * 3,
+        }}
+      >
+        {label}
+      </div>
+    </div>
+  );
 }
 
-export const isNumericalString = (value: string) => !isNaN(Number(value));
+function NumberIcon({ number }: { number: number }) {
+  return (
+    <EuiFlexGroup
+      justifyContent="spaceAround"
+      className="xyAnnotationNumberIcon"
+      gutterSize="none"
+      alignItems="center"
+    >
+      <EuiText color="ghost" className="xyAnnotationNumberIcon__text">
+        {number < 10 ? number : `9+`}
+      </EuiText>
+    </EuiFlexGroup>
+  );
+}
+
+const isNumericalString = (value: string) => !isNaN(Number(value));
+
+export const AnnotationIcon = ({
+  type,
+  rotateClassName = '',
+  isHorizontal,
+  renderedInChart,
+  ...rest
+}: {
+  type: string;
+  rotateClassName?: string;
+  isHorizontal?: boolean;
+  renderedInChart?: boolean;
+} & EuiIconProps) => {
+  if (isNumericalString(type)) {
+    return <NumberIcon number={Number(type)} />;
+  }
+  const iconConfig = annotationsIconSet.find((i) => i.value === type);
+  if (!iconConfig) {
+    return null;
+  }
+  return (
+    <EuiIcon
+      {...rest}
+      type={iconConfig.icon || type}
+      className={classnames(
+        { [rotateClassName]: iconConfig.shouldRotate },
+        {
+          lensAnnotationIconFill: renderedInChart && iconConfig.canFill,
+        }
+      )}
+    />
+  );
+};
+
+interface MarkerConfig {
+  axisMode?: YAxisMode;
+  icon?: string;
+  textVisibility?: boolean;
+  iconPosition?: IconPosition;
+}
+
+export function Marker({
+  config,
+  isHorizontal,
+  hasReducedPadding,
+  label,
+  rotateClassName,
+}: {
+  config: MarkerConfig;
+  isHorizontal: boolean;
+  hasReducedPadding: boolean;
+  label?: string;
+  rotateClassName?: string;
+}) {
+  if (hasIcon(config.icon)) {
+    return (
+      <AnnotationIcon type={config.icon} rotateClassName={rotateClassName} renderedInChart={true} />
+    );
+  }
+
+  // if there's some text, check whether to show it as marker, or just show some padding for the icon
+  if (config.textVisibility) {
+    if (hasReducedPadding) {
+      return <MarkerBody label={label} isHorizontal={isHorizontal} />;
+    }
+    return <EuiIcon type="empty" />;
+  }
+  return null;
+}
